refactor(transformers): rename addURLParams
