+++ src/transformers/index.js
@@ -3,7 +3,7 @@ const minify = require('./minify')
 const markdown = require('./markdown')
 const prettify = require('./prettify')
 const ensureSixHEX = require('./sixHEX')
-const addURLParams = require('./addURLParams')
+const addURLParams = require('./urlParams')
 const preventWidows = require('./preventWidows')
 const safeClassNames = require('./safeClassNames')
 const replaceStrings = require('./replaceStrings')
